:recycle: api: move CertificationChallengeLiveAlert imports from lib to src
