+++ api/lib/domain/models/index.js
@@ -42,7 +42,6 @@ import { CertificationCenterInvitation } from './CertificationCenterInvitation.j
 import { CertificationCenterInvitedUser } from './CertificationCenterInvitedUser.js';
 import { CertificationCenterMembership } from './CertificationCenterMembership.js';
 import { CertificationChallenge } from './CertificationChallenge.js';
-import { CertificationChallengeLiveAlert } from './CertificationChallengeLiveAlert.js';
 import { CertificationChallengeWithType } from './CertificationChallengeWithType.js';
 import { CertificationContract } from './CertificationContract.js';
 import { CertificationCourse } from './CertificationCourse.js';
@@ -197,7 +196,6 @@ export {
   CertificationCenterInvitedUser,
   CertificationCenterMembership,
   CertificationChallenge,
-  CertificationChallengeLiveAlert,
   CertificationChallengeWithType,
   CertificationContract,
   CertificationCourse,
