+++ api/src/certification/session/infrastructure/repositories/certification-challenge-live-alert-repository.js
@@ -1,6 +1,8 @@
 import { knex } from '../../../../../db/knex-database-connection.js';
-import { CertificationChallengeLiveAlert } from '../../../../../lib/domain/models/index.js';
-import { CertificationChallengeLiveAlertStatus } from '../../../../../lib/domain/models/CertificationChallengeLiveAlert.js';
+import {
+  CertificationChallengeLiveAlert,
+  CertificationChallengeLiveAlertStatus,
+} from '../../domain/models/CertificationChallengeLiveAlert.js';
 
 const save = async function ({ certificationChallengeLiveAlert }) {
   return knex('certification-challenge-live-alerts')
