+++ api/tests/shared/unit/domain/models/Assessment_test.js
@@ -1,6 +1,6 @@
 import { expect, domainBuilder } from '../../../../test-helper.js';
 import { Assessment } from '../../../../../src/shared/domain/models/Assessment.js';
-import { CertificationChallengeLiveAlertStatus } from '../../../../../lib/domain/models/CertificationChallengeLiveAlert.js';
+import { CertificationChallengeLiveAlertStatus } from '../../../../../src/certification/session/domain/models/CertificationChallengeLiveAlert.js';
 
 describe('Unit | Domain | Models | Assessment', function () {
   describe('#constuctor', function () {
